+++ src/adaptors/deepdao.ts
@@ -1,6 +1,7 @@
 import { fetcher } from '../utils';
 
 interface DeepdaoResp {
+  message: string;
   data: {
     score: number;
     rank: number;
@@ -22,14 +23,16 @@ export default async function getDeepDaoData(address: string) {
     }
   )) as DeepdaoResp;
 
-  const resp = {
-    score: json.data.score,
-    rank: json.data.rank,
-    relativeScore: json.data.relativeScore,
-    daos: json.data.daos,
-    proposals: json.data.proposals,
-    votes: json.data.votes,
-  };
-
-  return resp;
+  if ('message' in json) {
+    return {};
+  } else {
+    return {
+      score: json.data?.score,
+      rank: json.data?.rank,
+      relativeScore: json.data?.relativeScore,
+      daos: json.data?.daos,
+      proposals: json.data?.proposals,
+      votes: json.data?.votes,
+    };
+  }
 }