+++ src/adaptors/index.ts
@@ -22,6 +22,7 @@ export { default as getHiveOneData } from './hiveone';
 export { default as checkIdena } from './idena';
 export { default as getKarmaData } from './karma';
 export { default as getKnownOriginData } from './knownorigin';
+export { default as getLayer3Data } from './layer3';
 export { default as getLensData } from './lens';
 export { default as getMetagameData } from './metagame';
 export { default as getMewData } from './mew';